requirements.txt
aiosmtpd
argon2-cffi